[Version] Refactor version updating logic
